remotecommand: treat EOF as an error for v4+
Version 4 (and newer) of the remote command protocol specify that it is
expected to get a metav1.Status response after termination of the remote
process. However, the client-go implementation of the client side of
this protocol considered an EOF of the error stream (on which this
response is expected) to be a successful termination of
StreamWithContext.

This patch changes that behaviour for v4 and up, by specifying whether
an empty message is to be considered a success.

Notably, this fixes the case where StreamWithContext could return
successfully due to network blips severing the connection, and hence
causing an EOF of the error stream, which has caused flakes in cilium's
CI in the past.

Signed-off-by: David Bimmler <david.bimmler@isovalent.com>