remotecommand: tests: write status to stream
In v4, the protocol postulates that a metav1.Status is written to the
stream before termination. Some tests didn't do this. Luckily, there is
an existing test server fixture which handles this correctly, so let's
use that to fix the tests.

The main point of this change is to ensure that the tests are correct on
the protocol level, so that the change in the next patch (which makes
unexpected EOFs an error) can pass them without modification.

Signed-off-by: David Bimmler <david.bimmler@isovalent.com>