dra kubelet: parallelize unit test
The package contains two tests that time out after 45 seconds. When run
sequentially, testing the package already takes 1:30 minutes. Together with
other delays in the CI, the overall duration is approaching the timeout of 3
minutes, which would start to cause random test failures.

Parallelizing everything ensures that we use as little wall clock time as
possible.