Fix printers tests - remove dependency on leap years
Test that were using AddDate(+y, 0, 0) and then time.Sub were
sensitive to a specific date of their execution.

An example is a test with AddDate(-5, 0, 0) when executed
on 28th of February 2024 and when executed on 1st of March 2024.

The difference seen by Sub will be 5y1d in the first case
and 5y2d in the second case, because 29th of Feb 2024 is a leap
year as well as 29th of Feb 2020 that falls within the 5 year
difference.

Signed-off-by: Martin Sivak <msivak@redhat.com>