kubelet: check that the pod is finished before removing the container.
When pods with ephemeral storage limits exceed their capacity, the
kubelet tries to evict and kill the pod gracefully. We noticed that the
pod status changes to ContainerStatusUnknown. This is due to a race
condition removing the container before pod status can be updated.

In this change, we try to ensure that the pod is finished before
removing the container.