Continue streaming kubelet logs when runtime is unavailable
Container runtimes are able to run existing containers even when their
main CRI server is not available for any reason. The call to the
container status RPC happens quite frequently during log parsing, means
that a single CRI interruption will also abort streaming the logs.

We now check that specific use case and continue following the log
streaming if the CRI is unavailable. We still abort the streaming
accordingly if the CRI comes back and the container status reports that
the workload has exited.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>