Allow to build with go 1.21
The glibc bug used by go 1.22 is breaking runc 1.2.0 in some cases. This
does not directly influence Kubernetes, but downstream may rely on
Kubernetes as well as runc in their build environments. This means that
forcing users to build Kubernetes with go 1.22 may break their build
toolchain, which is especially the case for container runtimes. We now
relax that restriction to allow building with go 1.21 until the runc
issue has been resolved.

Refers to: https://github.com/opencontainers/runc/issues/4233

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>