e2e node: fix -v support
Since 43539c855f59f3318de73602d091c62ee5863cda (first released in
v1.30.0-alpha.2), the test/e2e/framework manages -v and -vmodule and uses them
for a logger which writes to the Ginkgo output stream.

This did not work for test/e2e_node, because:
- logs.AddFlags(pflag.CommandLine) registers its own -v and -vmodule flags
- pflag.CommandLine.AddGoFlagSet(flag.CommandLine) skips the corresponding
  flags in the flag.CommandLine
- pflag.Parse() initializes the settings in the "logs" package even though
  those are not used at runtime

The solution is to not use the "logs" package.