upgradeaware: avoid race condition in proxying
The upgradeaware proxy needs to avoid resource exhaustion in the case of
either side hanging up unexpectedly, c.f. 28376c70321 (Immediate close
the other half of the connection when proxying).

However, in the case of a network blip where the connection between
client and apiserver stalls, but the connection to the backend is
intact, the above fix can lead to a race condition. Specifically, I've
witnessed this occurring in the case of SPDY connections (for
remotecommand). On the SPDY level, there is an idle timeout mechanism,
which causes the _client_ to close their side of the connection. Once
this close makes it to the proxy, it immediately proceeds to close the
'writerComplete' channel, and (deferred) also the proxy's side of the
client connection. This is racy, as there may still be data being copied
towards the client at this point.

I don't know how we can avoid this race in general while also keeping
the resource exhaustion prevention in mind, but the proposed fix would
at least give the proxy a bit of time to copy in-flight data.

Note that the race could in theory also occur in the symmetric case of
network failure towards the backend, hence we need to also check for
that case.

The proposed fix works by setting read/write deadlines on the
connections involved in the `io.Copy` we are waiting for, hence limiting
the amount of time the connections are kept alive. The deadline is set
at half a second in the future, as suggested by aojea - the same as
socat uses by default.

Signed-off-by: David Bimmler <david.bimmler@isovalent.com>