apiserver/endpoints/installer: use storage GV for hubGV
Strategic merge patch assumes resources should be converted to
an internal resource version, presumably for etcd internal storage. It
seems the example apiservice gets around this by registering versions
with the internal version but this causes problems with much of the
other tooling that presumes resources have a version. At any rate, it
seems reasonable to allow an external apiservice to be able to choose
its own hubGroupVersion while still being able to utilize the
heavy-lifting logic in endpoints/installer.go

To achieve this, allow the rest.Storage parameter to drive what the
hubGroupVersion is when it implemnets `rest.StorageVersionProvider`. The
assumption here is that a hubVersion is always the same as the
storageVersion. (If this is not true, corrections wanted).

Signed-off-by: Andrew DeMaria <ademaria@cloudflare.com>