ConfigMaps: Manage Generation
This change makes the apiserver manage the Generation on ConfigMaps by
setting it to `1` on creation and incrementing it whenever the `Data` or
`BinaryData` changes. Changes to the object as a whole can already be
deducted from the `ResourceVersion` changing.

The main use-case is controllers that react to `ConfigMap` content
changes, with this change they can get away with only watching the
ObjectMeta.