Secrets: Manage Generation
This change makes the apiserver manage the Generation on Secrets by
setting it to `1` on creation and incrementing it whenever the `Data`
changes. Changes to the object as a whole can already bededucted
from the `ResourceVersion` changing.

The main use-case is controllers that react to `Secret` content
changes, with this change they can get away with only watching the
ObjectMeta