Disable P&F by default for aggregated apiservers
Priority and Fairness is currently enabled by default for any consumers
of the apiserver library. This include the kube-apiserver as well as
aggregated apiservers that are based on the generic apiserver library.
For the latter it doesn't really make sense to have P&F enabled by
default because the apiservers are for the majority not supposed to be
accessed directly, but rather via the kube-apiserver. In case there is
still a need to enable P&F at an aggregated apiserver level, they can
continue to do it either via the `--enable-priority-and-fairness` CLI
flag or by turning on the feature like we do for the kube-apiserver.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>