lease controller: add lease renew policies
Lease renew policies can be used to specify how the lease renew interval
is decided.

The previous default policy is using a fixed renew interval, which
updates lease for every 1/4 `nodeLeaseDurationSeconds` or
`renewInterval` provided in NewController().

A new policy is added in this commit to have dynamic renew intervals,
which starts with 1/2 `nodeLeaseDurationSeconds` and decreases 50%
of previous interval for next retry. It decreases 2 times and reset
the interval back to the initial.

A feature gate "DynamicNodeLeaseRenewInterval" is added and enabled by
default to apply new policy.

Signed-off-by: Eric Lin <exlin@google.com>