Add context-aware versions of retry.OnError and retry.OnConflict
The current util methods for retry don't take in a context. This adds in
versions which do take a context and respect the context cancellation or
deadline. The implementation is rather simple, since the underlying
method used, wait.ExponentialBackoff already has a context-aware
version, wait.ExponentialBackoffWithContext.

The documentation updates follow how the http.Request package refers to
the context-aware version:

https://pkg.go.dev/net/http@go1.22.2#NewRequest

http.NewRequest refers to http.NewRequestWithContext and the latter
contains the fuller documentation.