kubeadm: fix upgrading external CA cluster to 1.29
If the user has deleted ca.key from disk, this means that kubeadm
cannot sign any new certificates on 'upgrade apply'. In 'upgrade apply'
for 1.29 a call to generate the separate 'super-admin.conf'
and apply the RBAC for the new 'admin.conf' was added. This breaks for
external CA user.

If external CA is detected, show a warning that the user
must perform manual steps and apply only the RBAC without
generating the new 'super-admin.conf' and 'admin.conf'.