client-go: add metric for waitingForQueue depth
The workqueue package provides an abstraction for items to be added
after a delay. This is done by maintaining a min heap, called
the waitingForQueue that orders items by their delay time. A goroutine
polls this heap and feeds the items which are ready to be worked on
to the main queue.

This PR introduces a gauge metric for the waitingForQueue depth.
While a metric already exists for the depth of the main queue,
it doesn't provide observability to track how many items will be
enqueued in the future. Adding this metric would help downstream
users of the library, such as controller-runtime, with a large
waitingForQueue depth and a low main queue depth help alert on
long delays due to aggressive rate limiting.