Log computed QoS on error instead of status QoS
Pod QoS is immutable, but it is a computed field, so in-place resource
changes can result in updates being rejected because they would change
the QoS class -- e.g. removing all resources from a Burstable pod would
re-compute its QoS class as BestEffort, so it's not allowed.

When we return an error for a re-computed QoS, we erroneously log the
existing/status QoS class rather than the new computed one, which makes
it very confusing, e.g. a Burstable pod getting re-calculated as
BestEffort would log that Qos could not be changed to Burstable because
it was immutable.

This commit just changes the output such that it properly logs the
computed QoS that couldn't be applied, which should make it easier for
the user to understand why their resource changes are impossible.

Signed-off-by: John Kyros <jkyros@redhat.com>