garbagecollector: fix logging of failed discoveries
In GetDeletableResources() if we encounter a lookup error while
gathering preferred resources, the log message is e.g.:
`internal error: json: unsupported type: map[schema.GroupVersion]error`

encoding/json can't handle a map with a struct type as a key.  It also
won't convert errors to strings.

This commit converts the map to map[string]string and uses it for the
log message.