Define a common Node autoscaling safe-to-evict/do-not-disrupt annotation
Currently, there are 2 Node autoscalers sponsored by sig-autoscaling,
each supporting a different Pod annotation with the same semantics:

* Cluster Autoscaler:
  cluster-autoscaler.kubernetes.io/safe-to-evict=true/false
* Karpenter: karpenter.sh/do-not-disrupt=true

The semantics for cluster-autoscaler.kubernetes.io/safe-to-evict=false,
and karpenter.sh/do-not-disrupt=true are identical. Both of these
annotations will be replaced by
node-autoscaling.kubernetes.io/safe-to-evict=false.

cluster-autoscaler.kubernetes.io/safe-to-evict=true doesn't have an
equivalent in Karpenter right now, as Karpenter doesn't have any
pod-level conditions blocking consolidation. This means that the
equivalent new annotation
node-autoscaling.kubernetes.io/safe-to-evict=true should be trivially
supported by Karpenter initially (but will require caution if Karpenter
ever adds any pod-level conditions blocking consolidation).

Going with the Cluster Autoscaler wording for the common annotation, as
otherwise we'd have a double negation (do-not-disrupt=false) in the
safe-to-evict=true case which doesn't seem ideal.

This is a part of a broader alignment between Cluster Autoscaler
and Karpenter. More details about the alignment can be found in
https://docs.google.com/document/d/1rHhltfLV5V1kcnKr_mKRKDC4ZFPYGP4Tde2Zy-LE72w