cacher: short circuit conditional progress notifier
There can exist an extremely unlikely scenario where
the progress requester can repeatedly request progress
notifications even when the feature might not be supported.

The feature checker returns `true` if any node in an etcd
cluster has the feature enabled. However, it may change this
to `false` if it eventually detects a node which does not
have this enabled. Ideally, till we detect `false` we work
under the assumption that this feature is supported, however
we should stop requesting progress notifications as soon as
we get a `false` in order to prevent loading etcd with
redundant requests.

Signed-off-by: Madhav Jivrajani <madhav.jiv@gmail.com>