client-go: only filter PATCH events by default
PR #47367 introduced client side event spam filtering to reduce the API
server/etcd pressure of processing repeated events.  However, it use only
Source and InvolvedObject as key, so it can also drop important and unique
events. Slowing down the diagnosing significantly.

PR #103918 try to resolve this by adding a new SpamKeyFunc to customize the key
used. But this is still not ideal, because every user needs to dig into the
code of client-go to figure out why his event does not appear. If the
EventBroadcaster is initialized in a library, it may be even harder to set the
SpamKeyFunc.

I propose to only drop PATCH events by default, which is the intent of the
original proposal.  For the users who don't set SpamKeyFunc, they should get
exactly more events.