client-go record: avoid panic when watch creation failed
The previous attempt to fix this in
https://github.com/kubernetes/kubernetes/commit/6aa779f4ed3d3acdad2f2bf17fb27e11e23aabe4#diff-efa2cd1347df22ace5a516ea794152d00ef2a079db135c81787ed920ecb73658
didn't address the root cause (or perhaps created it, not sure): the goroutine
must not be started if watch creation failed.

Instead, the error gets logged (as before) and an empty watch gets returned to
the caller (new). This is necessary because the function doesn't have an error
return value and changing that now would be disruptive. The empty watch is
valid and usable, so callers won't crash when they calls Stop.

This showed up recently in failed unit tests, probably because test
cancellation makes this error more likely:

   "Unable start event watcher (will not retry!)" err="broadcaster already
   stopped" logger="TestGarbageCollectorConstruction leaked goroutine"

The logger value and a preceding warning show that this occurs after test
completion.