PSA: relax validation for capabilities for baseline + not hostUsers
pods having access to the `unshare` syscall (which is only blocked by seccomp) means
they have access to all capabilities, within a user namespace. Running `unshare -U` within
a pod does this.

Since the baseline policy doesn't force a seccomp profile, a pod user functionally can access
the same security profile as allowing all capabilities in the pod if the pod is in a user namespace.

Thus, we can safely relax the baseline profile to allow this explicitly.

Signed-off-by: Peter Hunt <pehunt@redhat.com>